refactor(pii): move PII detection to text/tasks/pii_detection
Pure move, no behaviour change. The input is a transcript, so the module
belongs under text/. workflows/audio_analysis/pii.py is a temporary re-export
shim, replaced by a real adapter later in this plan.

Also moves pii_test.py / pii_subprocess_test.py alongside the code they
exercise (they patch module globals the relocated code actually calls at
runtime; the compatibility shim's own globals are not on that call path,
so leaving the tests at the old import would silently stop mocking the
subprocess and hit the real one).
